Newspaper obstacle and colliders.
Re-using code from the chair moving and dirt spawning I have made a cube with the newspaper asset applied (with rotation) spawn.

Upon interaction, it activates an image of the newspaper asset, you can currently click and drag (touch hold and swipe) it out of the way, but it will disappear after 1.5 seconds as long as you have interacted with it. (Probably needs better design). When disappearing it just sets it to false and resets its position and timer, re-using the same canvas. It also disappears on death.

I have also slightly changed how colliders work. Previously, any collider that is marked with "is trigger" and interacts with the player would kill it. Now I have made it so only objects with the "KillingObstacle" tag will kill the player, and the tag is applied to chairs and birds. Now things like newspapers can collide and cause interactions without killing the player.

I wanted to make it so you can't move the player while the newspaper is on the screen but I struggled on how to disable the movement script in time. Can probably be easily done by finding the player and temporarily disabling the movement scripts while the paper is active.